fix/imageLoader directive
