+++ src/components/Logo.vue
@@ -1,6 +1,6 @@
 <template>
   <a href="#" class="logo">
-    <img class="logo-image" :src="require(`@/assets/logo.png`)" alt="Netflix roulette" />
+    <img v-lazyLoading class="logo-image" :data-src="require(`@/assets/logo.png`)" alt="Netflix roulette" />
   </a>
 </template>
