+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ImageWithLoader } from './ImageWithLoader.vue';
+export { default as Header } from './Header.vue';
+export { default as LocalLoader } from './LocalLoader.vue';
