+++ src/plugins/index.ts
@@ -1,0 +1,2 @@
+export { default as LazyLoadingPlugin } from './LazyLoadingPlugin';
+export { default as ScrollIntoView } from './ScrollIntoView';
